reducers: extract flashReducer to its own file
